requirements.txt
# Core ML
xgboost>=2.0.0
scikit-learn>=1.4.0
shap>=0.45.0
joblib>=1.3.0

# Data
pandas>=2.1.0
numpy>=1.26.0
pyyaml>=6.0.1

# Routing
networkx>=3.2.0

# API
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
python-dotenv>=1.0.0
httpx>=0.27.0

# Google Maps (optional — haversine fallback if not configured)
googlemaps>=4.10.0

# Dashboard
streamlit>=1.33.0
plotly>=5.20.0

# Optional: live weather
requests>=2.31.0

# Notebooks
jupyter>=1.0.0
matplotlib>=3.8.0
